+++ infrastructures/doctrine/DBSource/ODM/BatchManipulationManager.php
@@ -32,8 +32,6 @@ use Teknoo\East\Common\Contracts\DBSource\ManagerInterface;
 use Teknoo\East\Common\Contracts\Object\ObjectInterface;
 use Teknoo\East\Common\Contracts\Query\DeletingQueryInterface;
 use Teknoo\East\Common\Contracts\Query\UpdatingQueryInterface;
-use Teknoo\East\Common\DBSource\Manager\AlreadyStartedBatchException;
-use Teknoo\East\Common\DBSource\Manager\NonStartedBatchException;
 use Teknoo\Recipe\Promise\PromiseInterface;
 
 /**
@@ -114,6 +112,27 @@ class BatchManipulationManager implements BatchManipulationManagerInterface
     public function flush(): ManagerInterface
     {
         $this->baseManager->flush();
+
+        return $this;
+    }
+
+    public function registerFilter(string $className, array $parameters = [], bool $enabling = true): ManagerInterface
+    {
+        $this->baseManager->registerFilter($className, $parameters, $enabling);
+
+        return $this;
+    }
+
+    public function enableFilter(string $className): ManagerInterface
+    {
+        $this->baseManager->enableFilter($className);
+
+        return $this;
+    }
+
+    public function disableFilter(string $className): ManagerInterface
+    {
+        $this->baseManager->disableFilter($className);
 
         return $this;
     }